requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
certifi==2026.4.22
click==8.3.3
fastapi==0.136.1
GeoAlchemy2==0.19.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
packaging==26.2
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
geoalchemy2
shapely